Add support for base cases (and break lambdas)
