+++ snippets/base_case.lqt
@@ -4,6 +4,13 @@ let fib int -> int (
     dup 1 - fib swap 2 - fib +
 )
 
+let factorial 0 -> int (1)
+let factorial int -> int (
+    dup 1 - factorial *
+)
+
 let main -> (
-    5 fib print
+    [0 1 2 3 4 5 6 7 8 9 10]
+        (fib) map
+        print
 )
