+++ src/compiler.rs
@@ -1,6 +1,7 @@
 use crate::{
     diagnostic::{Diagnostic, Severity},
-    parser::{Function, Operand, Operation, OperationKind},
+    parser::{Function, Operand, OperationKind},
+    type_checker::{CheckedFunction, CheckedOpKind, CheckedOperation},
 };
 use std::{
     collections::HashMap,
@@ -24,7 +25,7 @@ impl Compiler {
         };
     }
 
-    pub fn compile(&mut self, ops: &Vec<Operation>) -> io::Result<()> {
+    pub fn compile(&mut self, ops: &Vec<CheckedOperation>) -> io::Result<()> {
         //Preamble
         self.out_file.write(b"#include <stdio.h>\n").unwrap();
         self.out_file.write(b"#include <stdint.h>\n").unwrap();
@@ -388,22 +389,22 @@ impl Compiler {
         )?;
 
         for op in ops {
-            self.emit(op)?;
+            self.emit(&op.kind)?;
         }
 
         Ok(())
     }
 
-    fn emit(&mut self, op: &Operation) -> io::Result<()> {
-        match &op.kind {
-            OperationKind::Push { operand } => self.emit_push_operand(operand)?,
-            OperationKind::Pop => {
+    fn emit(&mut self, op_kind: &CheckedOpKind) -> io::Result<()> {
+        match op_kind {
+            CheckedOpKind::Push { operand } => self.emit_push_operand(operand)?,
+            CheckedOpKind::Pop => {
                 self.out_file.write(b"pop();\n")?;
             }
-            OperationKind::Dup => {
+            CheckedOpKind::Dup => {
                 self.out_file.write(b"dup();\n")?;
             }
-            OperationKind::Rot => {
+            CheckedOpKind::Rot => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value c = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -413,7 +414,7 @@ impl Compiler {
                 self.out_file.write(b"        push(a);\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Swap => {
+            CheckedOpKind::Swap => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -421,7 +422,7 @@ impl Compiler {
                 self.out_file.write(b"        push(b);\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Over => {
+            CheckedOpKind::Over => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -431,7 +432,7 @@ impl Compiler {
                 self.out_file.write(b"        push(_a);\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Add => {
+            CheckedOpKind::Add => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -439,7 +440,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 1, b.value + a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Sub => {
+            CheckedOpKind::Sub => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -447,7 +448,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 1, b.value - a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Mul => {
+            CheckedOpKind::Mul => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -455,7 +456,7 @@ impl Compiler {
                     .write(b"        push((Value) { 1, b.value * a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Div => {
+            CheckedOpKind::Div => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -463,7 +464,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 1, b.value / a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Mod => {
+            CheckedOpKind::Mod => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -471,7 +472,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 1, b.value % a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Eq => {
+            CheckedOpKind::Eq => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -479,7 +480,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 0, b.value == a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Lt => {
+            CheckedOpKind::Lt => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -487,7 +488,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 0, b.value < a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::LtEq => {
+            CheckedOpKind::LtEq => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -495,7 +496,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 0, b.value <= a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Gt => {
+            CheckedOpKind::Gt => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -503,7 +504,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 0, b.value > a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::GtEq => {
+            CheckedOpKind::GtEq => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -511,7 +512,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 0, b.value >= a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::LogicalAnd => {
+            CheckedOpKind::LogicalAnd => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -519,7 +520,7 @@ impl Compiler {
                     .write(b"        push((Value)  { 0, b.value & a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::LogicalOr => {
+            CheckedOpKind::LogicalOr => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file.write(b"        Value b = pop();\n")?;
@@ -527,51 +528,51 @@ impl Compiler {
                     .write(b"        push((Value)  { 0, b.value | a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::LogicalNot => {
+            CheckedOpKind::LogicalNot => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"        Value a = pop();\n")?;
                 self.out_file
                     .write(b"        push((Value)  { 0, !a.value });\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Print => {
+            CheckedOpKind::Print => {
                 self.out_file.write(b"    print();\n")?;
             }
-            OperationKind::FunctionDefinition { function } => {
+            CheckedOpKind::FunctionDefinition { function } => {
                 self.emit_function_definition(function)?
             }
-            OperationKind::FunctionCall { name } => {
+            CheckedOpKind::FunctionCall { name } => {
                 write!(self.out_file, "    {}();\n", name)?;
             }
-            OperationKind::Map => {
+            CheckedOpKind::Map => {
                 self.out_file.write(b"    map();\n")?;
             }
-            OperationKind::Filter => {
+            CheckedOpKind::Filter => {
                 self.out_file.write(b"    filter();\n")?;
             }
-            OperationKind::Foreach => {
+            CheckedOpKind::Foreach => {
                 self.out_file.write(b"    foreach();\n")?;
             }
-            OperationKind::Fold => {
+            CheckedOpKind::Fold => {
                 self.out_file.write(b"    fold();\n")?;
             }
-            OperationKind::Split => {
+            CheckedOpKind::Split => {
                 self.out_file.write(b"    split();\n")?;
             }
-            OperationKind::Concat => {
+            CheckedOpKind::Concat => {
                 self.out_file.write(b"    concat();\n")?;
             }
-            OperationKind::Partial => {
+            CheckedOpKind::Partial => {
                 self.out_file.write(b"    partial();\n")?;
             }
-            OperationKind::Call => {
+            CheckedOpKind::Call => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"    Value lambda = pop();\n")?;
                 self.out_file
                     .write(b"        (((void (*)(void))lambda.value))();\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Len => {
+            CheckedOpKind::Len => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"    Value array = pop();\n")?;
                 self.out_file
@@ -582,7 +583,7 @@ impl Compiler {
                 )?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Pick => {
+            CheckedOpKind::Pick => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"    Value index = pop();\n")?;
                 self.out_file.write(b"    Value array = pop();\n")?;
@@ -594,7 +595,7 @@ impl Compiler {
                 self.out_file.write(b"        push(el);\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::Slice => {
+            CheckedOpKind::Slice => {
                 self.out_file.write(b"    {\n")?;
                 self.out_file.write(b"    Value upper = pop();\n")?;
                 self.out_file.write(b"    Value lower = pop();\n")?;
@@ -609,13 +610,13 @@ impl Compiler {
                 self.out_file.write(b"        push(array);\n")?;
                 self.out_file.write(b"    }\n")?;
             }
-            OperationKind::If => {
+            CheckedOpKind::If => {
                 self.out_file.write(b"    _if();\n")?;
             }
-            OperationKind::Return => {
+            CheckedOpKind::Return => {
                 self.out_file.write(b"    return;\n")?;
             }
-            OperationKind::Args => {
+            CheckedOpKind::Args => {
                 self.out_file.write(
                     b"{
                         Array *arr = (Array *)malloc(sizeof(Array));
@@ -632,7 +633,7 @@ impl Compiler {
                     }\n",
                 )?;
             }
-            OperationKind::Cons => {
+            CheckedOpKind::Cons => {
                 self.out_file.write(
                     b"{
                         Value a = pop();
@@ -651,7 +652,7 @@ impl Compiler {
                     }\n",
                 )?;
             }
-            OperationKind::Append => {
+            CheckedOpKind::Append => {
                 self.out_file.write(
                 b"{
                         Value array = pop();
@@ -668,7 +669,7 @@ impl Compiler {
                     }\n",
             )?;
             }
-            OperationKind::Range => {
+            CheckedOpKind::Range => {
                 self.out_file.write(
                     b"{
         Value upper = pop();
@@ -687,7 +688,8 @@ impl Compiler {
                     }\n",
                 )?;
             }
-            _ => todo!("{:?}", op),
+            CheckedOpKind::FunctionBaseCaseDefinition { .. } => {}
+            _ => todo!("{:?}", op_kind),
         }
         Ok(())
     }
@@ -754,47 +756,60 @@ impl Compiler {
         }
     }
 
-    fn predefine_lambdas(&mut self, ops: &Vec<Operation>) -> io::Result<()> {
-        for op in ops {
-            if let OperationKind::Push { operand } = &op.kind {
-                if let Operand::Seq { ops: lambda } = operand {
-                    self.predefine_lambdas(lambda)?;
-                    write!(
-                        self.out_file,
-                        "void lambda_{}() {{\n",
-                        self.next_lambda_label
-                    )?;
-                    if !self.lambdas.contains_key(operand) {
-                        self.lambdas.insert(
-                            operand.clone(),
-                            format!("lambda_{}", self.next_lambda_label),
-                        );
-                    }
-                    self.next_lambda_label += 1;
-                    for lambda_op in lambda {
-                        self.emit(&lambda_op)?;
-                    }
-                    self.out_file.write(b"}\n")?;
-                }
-            }
-        }
+    fn predefine_lambdas(&mut self, ops: &Vec<CheckedOperation>) -> io::Result<()> {
+        // for op in ops {
+        //     if let CheckedOpKind::Push { operand } = &op.kind {
+        //         //TODO tomorrow: checkedOperand
+        //         if let Operand::Seq { ops: lambda } = operand {
+        //             self.predefine_lambdas(lambda)?;
+        //             write!(
+        //                 self.out_file,
+        //                 "void lambda_{}() {{\n",
+        //                 self.next_lambda_label
+        //             )?;
+        //             if !self.lambdas.contains_key(&operand) {
+        //                 self.lambdas.insert(
+        //                     operand.clone(),
+        //                     format!("lambda_{}", self.next_lambda_label),
+        //                 );
+        //             }
+        //             self.next_lambda_label += 1;
+        //             for lambda_op in lambda {
+        //                 self.emit(&lambda_op.kind)?;
+        //             }
+        //             self.out_file.write(b"}\n")?;
+        //         }
+        //     }
+        // }
         Ok(())
     }
 
-    fn emit_function_definition(&mut self, function: &Function) -> io::Result<()> {
-        if let Operand::Seq { ops } = &function.body {
-            self.predefine_lambdas(ops)?;
+    fn emit_function_definition(&mut self, function: &CheckedFunction) -> io::Result<()> {
+        self.predefine_lambdas(&function.body)?;
 
-            if function.name == "main" {
-                write!(self.out_file, "void main(int argc, char *argv[]) {{\n")?;
-            } else {
-                write!(self.out_file, "void {}() {{\n", function.name)?;
-            }
-            for op in ops {
-                self.emit(&op)?;
-            }
+        if function.name == "main" {
+            write!(self.out_file, "void main(int argc, char *argv[]) {{\n")?;
         } else {
-            unreachable!()
+            write!(self.out_file, "void {}() {{\n", function.name)?;
+        }
+        for base_case in &function.base_cases {
+            write!(self.out_file, "{{\n")?;
+            write!(self.out_file, "Value input = stack[sp - 1];\n")?;
+            write!(
+                self.out_file,
+                "if (input.value == {}) {{\n",
+                base_case.input
+            )?;
+            write!(self.out_file, "pop();\n")?;
+            for op in &base_case.body {
+                self.emit(&op.kind)?;
+            }
+            write!(self.out_file, "return;\n")?;
+            write!(self.out_file, "}}\n")?;
+            write!(self.out_file, "}}\n")?;
+        }
+        for op in &function.body {
+            self.emit(&op.kind)?;
         }
         self.out_file.write(b"}\n")?;
         Ok(())
